TS-42: review and import Bryan Call's patches
git-svn-id: https://svn.apache.org/repos/asf/incubator/trafficserver/traffic/trunk@905704 13f79535-47bb-0310-9956-ffa450edef68

+++ iocore/cluster/ClusterCache.cc
@@ -2225,7 +2225,7 @@ CacheContinuation::remoteOpEvent(int event_code, Event * e)
     start_time = 0;
   }
   // for CACHE_EVENT_RESPONSE/XXX the lock was acquired at the higher level
-  int return_error = 0;
+  intptr_t return_error = 0;
   ClusterVCToken *pToken = NULL;
 
 retry:
